Add: shared pageinfo 엔티티 추가
