+++ src/schema.gql
@@ -2,6 +2,21 @@
 # THIS FILE WAS AUTOMATICALLY GENERATED (DO NOT MODIFY)
 # ------------------------------------------------------
 
+"""페이지네이션 정보, """
+type PageInfo {
+  """총 문서 수"""
+  total_count: Float!
+
+  """현재 페이지가 마지막 페이지인지 여부, 값이 false면 page를 증가시켜 다음 페이지를 요청할 수 있음"""
+  is_end: Boolean!
+
+  """노출 가능 페이지 수 (1 ~ 45)"""
+  pageable_count: Float!
+
+  """현재 페이지 번호"""
+  cur_page: Float!
+}
+
 """카카오 지도 api로 부터 받은 위치 정보"""
 type Place {
   """kakao place id"""
@@ -17,21 +32,6 @@ type Place {
   distance: String
   x: Float
   y: Float
-}
-
-"""place 페이지네이션 정보, """
-type PageInfo {
-  """검색어에 검색된 문서 수"""
-  total_count: Float!
-
-  """현재 페이지가 마지막 페이지인지 여부, 값이 false면 page를 증가시켜 다음 페이지를 요청할 수 있음"""
-  is_end: Boolean!
-
-  """노출 가능 페이지 수 (1 ~ 45)"""
-  total_page_count: Float!
-
-  """현재 페이지 번호"""
-  cur_page: Float!
 }
 
 """페이지네이션 정보를 포함한 place 정보"""
